--- lib/features/visit_tracker/acitivity/data/models/activity.dart
@@ -1,16 +1,0 @@
-import 'package:freezed_annotation/freezed_annotation.dart';
-
-part 'activity.freezed.dart';
-part 'activity.g.dart';
-
-@freezed
-abstract class Activity with _$Activity {
-  factory Activity({
-    required int id,
-    required String description,
-    @JsonKey(name: 'created_at') required String createdAt,
-  }) = _Activity;
-
-  factory Activity.fromJson(Map<String, dynamic> json) =>
-      _$ActivityFromJson(json);
-}
