+++ lib/features/visit_tracker/visits/data/models/visit.dart
@@ -17,3 +17,27 @@ abstract class Visit with _$Visit {
   }) = _Visit;
   factory Visit.fromJson(Map<String, dynamic> json) => _$VisitFromJson(json);
 }
+
+@freezed
+abstract class Activity with _$Activity {
+  factory Activity({
+    required int id,
+    required String description,
+    @JsonKey(name: 'created_at') required String createdAt,
+  }) = _Activity;
+
+  factory Activity.fromJson(Map<String, dynamic> json) =>
+      _$ActivityFromJson(json);
+}
+
+@freezed
+abstract class CustomerItem with _$CustomerItem {
+  factory CustomerItem({
+    required int id,
+    required String name,
+    @JsonKey(name: 'created_at') required String createdAt,
+  }) = _CustomerItem;
+
+  factory CustomerItem.fromJson(Map<String, dynamic> json) =>
+      _$CustomerItemFromJson(json);
+}
